requirements.txt
# Django Framework
django~=5.2.0
gunicorn~=23.0.0
django-environ~=0.12.0
psycopg2-binary~=2.9.10

# Frontend Integration
django-tailwind~=3.8.0

# Image Processing
Pillow~=10.4.0

# GenAI
google-api-core~=2.25.1
google-genai~=1.24.0
